Diariesのアクションを修正した

+++ backend/db/migrate/20211110060309_create_diaries.rb
@@ -4,7 +4,7 @@ class CreateDiaries < ActiveRecord::Migration[6.1]
   def change
     create_table :diaries do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer :sleeping_hours
+      t.float :sleeping_hours
       t.string :feeling
       t.string :date
